+++ get_rect/core/src/se/chalmers/get_rect/game/IGameComponent.java
@@ -3,6 +3,6 @@ package se.chalmers.get_rect.game;
 import se.chalmers.get_rect.adapters.IGraphicsAdapter;
 
 public interface IGameComponent {
-    void update(long delta);
+    void update(double delta);
     void draw(IGraphicsAdapter graphics);
 }
