+++ get_rect/core/src/se/chalmers/get_rect/game/screens/StartMenuScreen.java
@@ -19,7 +19,7 @@ public class StartMenuScreen implements IScreen {
     }
 
     @Override
-    public void update(long delta) {
+    public void update(double delta) {
 
     }
